requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
chromadb
pypdf
python-dotenv
langchain_chroma
sentence-transformers
pymupdf
rapidocr-onnxruntime
streamlit